+++ docs/source/api/algorithms/xgi.algorithms.properties.rst
@@ -1,0 +1,19 @@
+xgi.algorithms.properties
+=========================
+
+.. currentmodule:: xgi.algorithms.properties
+
+.. automodule:: xgi.algorithms.properties
+
+   .. rubric:: Functions
+
+   .. autofunction:: degree_counts
+   .. autofunction:: degree_histogram
+   .. autofunction:: density
+   .. autofunction:: edge_neighborhood
+   .. autofunction:: incidence_density
+   .. autofunction:: is_possible_order
+   .. autofunction:: is_uniform
+   .. autofunction:: max_edge_order
+   .. autofunction:: num_edges_order
+   .. autofunction:: unique_edge_sizes
